requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
backrefs==5.9
bcrypt==4.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
cloudpathlib==0.22.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
crewai==0.148.0
crewai-tools==0.0.1
cryptography==45.0.5
cymem==2.0.11
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
durationpy==0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
ghp-import==2.1.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
groq==0.30.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.9.2
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain-core==0.3.75
langchain-groq==0.3.6
langcodes==3.5.0
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.6
language_data==1.3.0
litellm==1.72.6
marisa-trie==1.3.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
networkx==3.5
nodeenv==1.9.1
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.97.0
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
paginate==0.5.7
pandas==2.3.2
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
posthog==5.4.0
pre_commit==4.2.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16
pymongo==4.13.2
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
RapidFuzz==3.14.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
tavily-python==0.7.9
tenacity==9.1.2
thinc==8.3.6
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.7.21
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0

whoosh==2.7.4

trafilatura
readability-lxml

scikit-learn
PyYAML
fastapi
uvicorn
dvc
beautifulsoup4
faiss-cpu
sentence-transformers
whoosh
